+++ ci-droid-tasks-consumer-services/src/main/java/com/societegenerale/cidroid/tasks/consumer/services/PushEventOnDefaultBranchService.java
@@ -11,8 +11,8 @@ import org.apache.commons.lang3.time.StopWatch;
 
 import java.util.List;
 
-import static com.societegenerale.cidroid.tasks.consumer.services.MonitoringAttributes.REPO;
-import static com.societegenerale.cidroid.tasks.consumer.services.MonitoringEvents.PUSH_EVENT_TO_PROCESS;
+import static com.societegenerale.cidroid.tasks.consumer.services.monitoring.MonitoringAttributes.REPO;
+import static com.societegenerale.cidroid.tasks.consumer.services.monitoring.MonitoringEvents.PUSH_EVENT_TO_PROCESS;
 import static java.util.stream.Collectors.toList;
 
 @Slf4j
