+++ ci-droid-tasks-consumer-services/src/main/java/com/societegenerale/cidroid/tasks/consumer/services/monitoring/MonitoringAttributes.java
@@ -1,14 +1,14 @@
-package com.societegenerale.cidroid.tasks.consumer.services;
+package com.societegenerale.cidroid.tasks.consumer.services.monitoring;
 
 public class MonitoringAttributes {
-
-    private MonitoringAttributes(){
-
-    }
 
     public final static String PR_NUMBER="prNumber";
 
     public final static String REPO="repo";
 
     public final static String DURATION="duration";
+
+    private MonitoringAttributes(){
+
+    }
 }
