+++ ci-droid-tasks-consumer-services/src/main/java/com/societegenerale/cidroid/tasks/consumer/services/monitoring/MonitoringEvents.java
@@ -1,10 +1,6 @@
-package com.societegenerale.cidroid.tasks.consumer.services;
+package com.societegenerale.cidroid.tasks.consumer.services.monitoring;
 
 public class MonitoringEvents {
-
-    private MonitoringEvents(){
-
-    }
 
     public final static String PUSH_EVENT_TO_PROCESS="pushEventToProcess";
 
@@ -21,4 +17,8 @@ public class MonitoringEvents {
     public final static String BULK_ACTION_PR_CREATED="bulkActionPrCreated";
 
     public final static String BULK_ACTION_PROCESSED="bulkActionProcessed";
+
+    private MonitoringEvents(){
+
+    }
 }
